bio printf: Safely range check real overflow on output.
There is a problem casting ULONG_MAX to double which gcc-10 is warning about.
ULONG_MAX typically cannot be exactly represented as a double.  ULONG_MAX + 1
can be and this fix uses the latter, however constructing this number cannot
be done using integer arithmetic.

The addresses a symptom of the underlying problem: we print doubles via an
unsigned long integer.  Doubles have a far great range and should be printed
better.