CONF: refactor NCONF_new_with_libctx() into separate functions
This removes NCONF_new_with_libctx() in favour of separate functions
for assigning a library context to the CONF object, and retrieve it.

The reason for refactoring is to give that other configuration loading
functions better flexibility in how they will combine diverse needs
and demands on the CONF object.