Visual Studio compiler switches:  Stack Probes, String Pooling
Summary:  /Gs0 is removed.  /GF is removed for debug but not for release.

Reason for not using /GF during debug bulid:  With this switch, it is impossible to step through code for "openssl s_server -WWW" due to significant code optimization.  Removing /GF fixes this problem (on Windows 10).

Reason for not using /Gs0:  With this switch, stepping into "openssl s_server -WWW" crashes at at least two places due to stack check failure.  For some reason, the behavior is avoided by switching to the dissassembly window, stepping into _stkchk, and then stepping through it until it returns.  Removing the switch fixes this problem.  It is removed for release as well as debug because Microsoft documentation states that the switch should not be used "unless you know exactly why you are using it".