TEST: Provide a default openssl.cnf for tests
In Debian the apps/openssl.cnf gets patched to provide a system wide
policy for TLSv1.2+ and SECLEVEL=2.
Somewhere between alpha1 and alpha3 the testsuite started failing
because the tests requiring < TLSv1.2 failed due to the restriction in
the modified .cnf file.
As far as I can tell, there is no need use the whole .cnf file from
apps which contains mainly certificate related options. Also the
testsuite should not depend on files outside of the testsuite.

Provide an empty .cnf file for the testsuite if none is specified. This
is enough to pass the testsuite. Rename the current "default" file to
"default_provider".

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>