Windows build: (1) string pooling only for release (2) default stack probes
Without these changes, Visual Studio's debugger cannot be used to step through code for "openssl s_server -WWW".  These changes fix that.  With string pooling, code is so optimized that execution stepping is confusing and ineffective.  The use of /Gs0 is contrary to Microsoft documentation, which states that nondefault stack probing should only be used if you know exactly why you need to do it.  Worse, /Gs0 appears to break Visual Studio's debugger.  (The workaround is to use the disassembler window and step into and through _chkstk.)