Always create $(DESTDIR)/lib during make install, even on Windows (#217)
Cross-compiling for Windows deposits binaries in bin, not lib.

Commit 72ea3e9 added rules to copy liboqs.a to OpenSSL's lib directory for development purposes, but when cross-compiling for Windows, lib hasn't been created yet.

As a result, during make install, liboqs.a ends up copied and named "lib" in $(DESTDIR). This later causes a failure when it tries to create a directory there called "lib" only to find a file exists with that name.

This commit causes lib to always be created in the install_runtime_libs recipe, and liboqs.a ends up in the right place.