ERR: refactor the treatment of system error messages
Recent experiences with GNU/hurd shows that we made assumption about
well known errno codes that aren't necessarily true.  GNU/hurd codes
appear to use the upper bits as error sections for diverse sub-systems.

We therefore need to rethink how we treat system error numbers
(errno), and pack them in such a way that they are recognised later.

To make that easier in the ERR sub-system, we add an internal function
err_pack(), which does special encoding of the reason code when
needed.  The macro ERR_PACK is thereby reserved for definition of
library specific codes and should never be used for system codes.

Fixes #12217
Fixes #12276