TEST: fix test/errtest.c
test/errtest.c used the system error code 1 for EPERM.  However, EPERM
may be coded differently on different systems, so we switch to using
EPERM instead.  To be safe (even though the error string for that code
is well defined in POSIX), we use strerror() to retrieve the string
for that error code.

Fixes #12276
Fixes #12217