Run tests in parallel
The environment variable `HARNESS_JOBS` can be used to control how many
jobs to run in parallel. It defaults to 4.

This commit does not define custom `rules`, and different versions of
`TAP::Harness` come with different strategies regarding the default
`rules` that define which test recipes can be run in parallel.
In recent versions of Perl, unless specified otherwise any task can be
run in parallel.