test/drbgtest: improve the reseed after fork test
Issue #12377 demonstrated that it is not sufficient to verify that
after a fork a reseeding is triggered in the child. This commit
enhances the test by collecting the output of the public and private
drbg for the parent and all children and checking for duplicates.
In case of duplicates, it displays a sorted output.

    *ERROR*: DUPLICATE RANDOM OUTPUT:
        output: bb836def, pid: 30831 (parent, public)
        output: bb836def, pid: 30837 (child 1, public)
        output: bb836def, pid: 30838 (child 2, public)
        output: bb836def, pid: 30839 (child 3, public)
        output: bb836def, pid: 30840 (child 4, public)
        output: bb836def, pid: 30841 (child 5, public)
        output: 02bf7135, pid: 30831 (parent, private)
        output: 02bf7135, pid: 30837 (child 1, private)
        output: 02bf7135, pid: 30838 (child 2, private)
        output: 02bf7135, pid: 30839 (child 3, private)
        output: 02bf7135, pid: 30840 (child 4, private)
        output: 02bf7135, pid: 30841 (child 5, private)