Configure: remove ref to missing Makefile.tmpl
This commit removes the reference to Makefile.tmpl in
Configurations/README.md, and also removes it from the
@build_file_templates array in the Configure script.

The motivation for removing it is that this file is referenced in
Configurations/README.md but as this file does not exist it leads to
a 404 Not Found.

I'm not really sure this should be removed from Configure, or if the
updates made to Configurations/README.md are correct but hopefully
others can chime in and provide advise on this.