Fix AES_SIV cipher dup_ctx and init.
Calling EVP_CIPHER_init multiple time resulted in a memory leak in the siv.
Fixing this leak also showed that the dup ctx was not working for siv mode.
Note: aes_siv_cleanup() can not be used by aes_siv_dupctx() as it clears data
that is required for the decrypt (e.g the tag).
This was found during testing of the init/update/init/update for all ciphers.